engine/execkit/atomic/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rpaatomic"
version = "1.1.82"
description = "rpaatomic."

requires-python = ">=3.7"
dependencies = [
    "PyYAML>=6.0.1",
    "rpaframe",
    "pycryptodome",
    "python-dateutil"
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true

[tool.setuptools]
package-dir = {"rpaatomic" = "src/rpaatomic"}
include-package-data = false

[tool.setuptools.packages.find]
where = ["src"]
include = ["rpaatomic*"]

[tool.setuptools.package-data]
rpaatomic = ["config.yaml"]